+++ egui/src/memory.rs
@@ -298,6 +298,7 @@ impl Focus {
         } else if self.pressed_tab && self.id == None && !self.give_to_next {
             // nothing has focus and the user pressed tab - give focus to the first widgets that wants it:
             self.id = Some(id);
+            self.pressed_tab = false;
         }
 
         self.last_interested = Some(id);
@@ -349,6 +350,11 @@ impl Memory {
     }
 
     /// Does this widget have keyboard focus?
+    ///
+    /// This function does not consider whether the UI as a whole (e.g. window)
+    /// has the keyboard focus. That makes this function suitable for deciding
+    /// widget state that should not be disrupted if the user moves away
+    /// from the window and back.
     #[inline(always)]
     pub fn has_focus(&self, id: Id) -> bool {
         self.interaction.focus.id == Some(id)