+++ egui_demo_lib/src/demo/misc_demo_window.rs
@@ -426,6 +426,7 @@ impl Tree {
             Tree(vec![Tree(vec![Tree::default(); 2]); 3]),
         ])
     }
+
     pub fn ui(&mut self, ui: &mut Ui) -> Action {
         self.ui_impl(ui, 0, "root")
     }
@@ -443,7 +444,7 @@ impl Tree {
     fn children_ui(&mut self, ui: &mut Ui, depth: usize) -> Action {
         if depth > 0
             && ui
-                .button(RichText::new("delete").color(Color32::RED))
+                .button(RichText::new("delete").color(ui.visuals().warn_fg_color))
                 .clicked()
         {
             return Action::Delete;