+++ emath/src/lib.rs
@@ -51,12 +51,14 @@ pub use {
 pub trait One {
     fn one() -> Self;
 }
+
 impl One for f32 {
     #[inline(always)]
     fn one() -> Self {
         1.0
     }
 }
+
 impl One for f64 {
     #[inline(always)]
     fn one() -> Self {
@@ -78,6 +80,7 @@ pub trait Real:
 }
 
 impl Real for f32 {}
+
 impl Real for f64 {}
 
 // ----------------------------------------------------------------------------
@@ -251,6 +254,7 @@ macro_rules! impl_num_ext {
             fn at_least(self, lower_limit: Self) -> Self {
                 self.max(lower_limit)
             }
+
             #[inline(always)]
             fn at_most(self, upper_limit: Self) -> Self {
                 self.min(upper_limit)
@@ -307,6 +311,28 @@ fn test_normalized_angle() {
     almost_eq!(normalized_angle(0.0), 0.0);
     almost_eq!(normalized_angle(TAU), 0.0);
     almost_eq!(normalized_angle(2.7 * TAU), -0.3 * TAU);
+}
+
+// ----------------------------------------------------------------------------
+
+/// Calculate a lerp-factor for exponential smoothing using a time step.
+///
+/// * `exponential_smooth_factor(0.90, 1.0, dt)`: reach 90% in 1.0 seconds
+/// * `exponential_smooth_factor(0.50, 0.2, dt)`: reach 50% in 0.2 seconds
+///
+/// Example:
+/// ```
+/// # use emath::{lerp, exponential_smooth_factor};
+/// # let (mut smoothed_value, target_value, dt) = (0.0_f32, 1.0_f32, 0.01_f32);
+/// let t = exponential_smooth_factor(0.90, 0.2, dt); // reach 90% in 0.2 seconds
+/// smoothed_value = lerp(smoothed_value..=target_value, t);
+/// ```
+pub fn exponential_smooth_factor(
+    reach_this_fraction: f32,
+    in_this_many_seconds: f32,
+    dt: f32,
+) -> f32 {
+    1.0 - (1.0 - reach_this_fraction).powf(dt / in_this_many_seconds)
 }
 
 // ----------------------------------------------------------------------------